==> What Beatles Song Is That?/01 - Pulling Lyrics Dataset Using Genius API and lyricsgenius.ipynb <==
## 1. Importing Python Libraries 

We shall start by importing the essential Python libraries

### IMPORTING LIBRARIES
import numpy as np
import pandas as pd
import lyricsgenius as lg

## 2. Connecting to the API

We now use the Client Access Token we had received when we signed up for Genius API and use the lyricsgenius library to gain access to the data available in genius.com. Here, we have few parameters to keep in mind:

1. timeout: this is the time in seconds before quitting the response
2. retries: the number of retries in case of timeouts and errors
3. remove_section_headers: to decide whether we want the lyrics to include headers defining sections as verses or choruses
4. skip_non_songs: to decide whether we want only songs or songs plus other contents like interviews
5. excluded_terms: helps to get rid of the remix, live and other versions of already existing songs by filtering out titles containing these terms

Here, we exclude the songs containing the terms _'Remix'_, _'Mix'_, _'Version'_ and _'Overdub'_ in their titles so that we can only get the original studio versions of the songs. So with these, we shall create an object _genius_ of class API.

### USING API TOKEN TO GET DATA
api_token = 'Client Token Access from Genius API'
genius = lg.Genius(api_token, timeout = 100, retries = 100, remove_section_headers = True, skip_non_songs = True, excluded_terms = ['(Remix)', '(Mix)', '(Version)', ('Overdub')])

## 3. Pulling the Data for Any Artist from the API

To store data properly and to make searches efficient, genius.com gives a unique _id_ to every artist. So, if we want to pull data for a specific artist, we shall first try to get their unique _id_. For this, we can use the _search_artists_ method in _genius_ followed by the name of any artist whose data we require. Here, we can try to find one for The Beatles.

### PULLING ARTIST DATA
artist_data = genius.search_artists('The Beatles')
print(artist_data)

If we look carefully, we see that the first value in the list that was returned has details for The Beatles, including their _'id'_ which is 586. With this unique _'id'_, we can now use the _search_artist_ method in _genius_ to get the data of all the songs recorded by The Beatles.

### PULLING DATA OF ALL SONGS PERFORMED BY THE ARTIST
artist = genius.search_artist('586', artist_id = 586, sort = 'title')

## 4. Data Exploration

Let us now take a look at the object that was returned.

print(artist)

So the object _artist_ contains data for 687 songs recorded by The Beatles. Keep in mind that the actual number of total songs recorded by them is only 213; thus we need to get rid of the extra songs which might be different versions of the same songs.

Now lets look at the song title and song lyrics contained in _artist_.

### PRINTING THE SONG TITLE
song = artist.songs[8].title
print(song)

We have our song title. Next, we have it's lyrics.

### PRINTING THE SONG LYRICS
lyrics = artist.songs[8].lyrics
print(lyrics)

Notice that the lyrics has the song title at the top followed by the word _'Lyrics'_. Also, it ends with the word _'Embed'_ along with a number. This is true for the lyrics of all the other songs. The formatting also needs to be dealt with since the lyrics contains a lot of _'\n'_ i.e. new line.

## 5. Cleaning the Lyrics Data I

Let us try to resolve these issues for this particular lyrics. To get rid of the song title on the first line, we split the lyrics using the word _'Lyrics'_ and only take the second portion. For the extra word in the last line, we again split the lyrics using the word _'Embed'_ and take only the first portion. Then in the remaining lyrics, we replace each _'\n'_ with a blank space. Let's find out if we solved these issues.

### CLEANING THE LYRICS OF A SINGLE SONG
lyrics1 = lyrics.split('Lyrics')[1]
lyrics2 = lyrics1.split('Embed')[0]
lyrics3 = lyrics2.replace('\n', ' ')
print(lyrics3)

The lyrics data looks cleaner than before although it still contains a numerical value at the end; this shall be resolved later. Now we need to do this for all the songs but before that we shall create a dataframe for the song titles and lyrics.

## 6. Creating a Lyrics Dataframe


First, let us create a list containing the song titles in one column and their respective lyrics in the other.

### CREATING A LIST 
lyrics_list = []
for song in range(0, len(artist)):
    song_title = artist.songs[song].title
    song_lyrics = artist.songs[song].lyrics
    lyrics_list.append([song_title, song_lyrics])

We shall now convert this list into a pandas dataframe.

### CREATING A PANDAS DATAFRAME
lyrics_df = pd.DataFrame(lyrics_list, columns = ['title', 'lyrics'])
lyrics_df.head(10)

Notice that the dataframe does not contain lyrics for every song title i.e. there are missing values in the data.

## 7. Cleaning the Lyrics Data II


Now, let us clean the lyrics of all the songs contained in the dataframe using the same process as above but this time in a loop.

### CLEANING THE LYRICS OF ALL THE SONGS
for i in range(0, len(lyrics_df)):
    if len(lyrics_df['lyrics'][i]) > 0:
        song_lyrics = lyrics_df['lyrics'][i].split('Lyrics')[1]
        song_lyrics1 = song_lyrics.split('Emded')[0]
        lyrics_df['lyrics'][i] = song_lyrics1.replace('\n', ' ')

We check to see if we were indeed able to clean the lyrics like we had wanted to.

lyrics_df.head(10)

## 8. Saving the Lyrics Dataframe

Now, let us save this dataframe in csv format which can be used later to make models.

lyrics_df.to_csv('beatles_lyrics_data.csv', index = False)
