==> What Beatles Song Is That?/02 - Pulling Song List From Wikipedia.ipynb <==
## 1. Importing Python Libraries

We shall start by importing the essential Python libraries.

### IMPORTING LIBRARIES
import numpy as np
import pandas as pd

## 2. Pulling Data from the URL 

Next, let us take the url of the wikipedia page that contains the table of all the songs recorded by The Beatles and import it using pandas.

### IMPORTING URL DATA
url = 'https://en.wikipedia.org/wiki/List_of_songs_recorded_by_the_Beatles'
song_list = pd.read_html(url)

Now, since the third table in the webpage contains the list of songs, let us make a pandas dataframe using the third table from the _song_list_ object.  

song_list = pd.DataFrame(song_list[2]) 

## 3. Data Exploration

Let us check the table.

song_list.head(10)

We see that although the dataframe contains the list of songs, all the song titles are written within inverted commas. Also, some of the titles have reference links in them. Further, the dataframe also contains a column which contains only reference numbers.

Next, let us check the dimensions of this dataframe.

song_list.shape

So, the dataframe contains the title of all the 213 songs recorded by The Beatles.

## 4. Cleaning the Titles Dataframe

Let us start cleaning the dataframe. First, we drop the column containing reference numbers.

### DROPPING UNNECESSARY COLUMNS
song_list.drop(['Ref(s)'], axis = 1, inplace = True)

Next, let us try to extract clean song titles. For this, we split each title using the punctuation _"_ and then take the second portion of each split.

### CLEANING THE SONG TITLES
title = []
for i in range(0, len(song_list)):
    song_title = song_list['Song'][i]
    song_title_clean = song_title.split('"')[1]
    title.append(song_title_clean)

Now, we replace the song title with the clean titles.

### REPLACING TITLES WITH CLEANED ONES
song_list['Song'] = title
song_list.head(10)

## 5. Saving the Titles Dataframe

Lastly, we save this dataframe in csv format.

song_list.to_csv('beatles_song_list.csv', index = False)
