==> What Beatles Song Is That?/03 - Getting Song Title From Lyrics.ipynb <==
## 1. Importing Python Libraries

We start by importing the essential Python libraries.

### IMPORTING LIBRARIES
import numpy as np
import pandas as pd
import nltk
import re
from nltk.corpus import stopwords
from nltk.stem.wordnet import WordNetLemmatizer
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.metrics import classification_report
from sklearn.naive_bayes import MultinomialNB
nltk.download('stopwords')
nltk.download('omw-1.4')

## 2. Data Importation

Now, we shall import the dataframe containing the lyrics for all the songs performed by The Beatles and display its first 10 rows.

### IMPORTING THE DATA
beatles_lyrics = pd.read_csv('beatles_lyrics_data.csv')
beatles_lyrics.head(10)

Next, we import the dataframe containing the list of all the unique songs recorded by The Beatles and display the first 10 rows.

beatles_songs = pd.read_csv('beatles_song_list.csv')
beatles_songs.head(10)

Now, what we'd like to do is merge these two dataframes. Since the dataframe _beatles_songs_ contains the unique titles of all the songs while _beatles_lyrics_ has lyrics for not only the official studio versions but also for different versions of the same songs, we would like to merge the two dataframes such that we get the lyrics for only those songs that appear in _beatles_songs_ dataframe. In doing so, we get rid of all the other versions and only have each song appear only once.

## 3. Cleaning the Song Titles

Before we go any further, let us make a function that will clean the title names in both the dataframes. We do this because we want to bring the song titles in both the dataframes down to the same format since we will be using these to merge the two dataframes. So we make a function _clean_title_ that will first turn each title to lower case then remove all punctuations and then get rid of all the white spaces to return us a clean title.  

### CREATING A FUNCTION TO CLEAN SONG TITLES
def clean_title(text):
    text = text.lower() #turns to lower case
    text = re.sub(r'[^\w\s]+', '', text) #removes punctuation
    text = text.strip() #removes white spaces
    return text

Next, we apply _clean_title_ to the columns that contain song titles in both the dataframes.

### APPLYING THE FUNCTION TO TITLES
beatles_lyrics['title'] = beatles_lyrics['title'].apply(clean_title)
beatles_songs['title'] = beatles_songs['Song'].apply(clean_title)

Let us check to see if it worked in the _beatles_lyrics_ dataframe.

beatles_lyrics.head(10)

We see that the song titles are all in lower case, without any punctuation and devoid of white spaces left behind after removal of punctuations.

We again check the same in _beatles_songs_ dataframe.

beatles_songs.head(10)

And, we see the same clean outcome.

## 4. Missing Value Treatment

Next, we shall deal with the missing values in the _beatles_lyrics_ dataframe by simply getting rid of all the rows that contain missing values.

### DELETING MISSING VALUES
beatles_lyrics.dropna(inplace = True)
beatles_lyrics.head(10)

So, we've removed all the missing values from _beatles_lyrics_ and now we are ready to merge the two datasets.

## 5. Merging the Two Dataframes

As mentioned earlier, we shall merge the two dataframes using the columns that contain the clean song titles.

### MERGING THE DATAFRAMES USING INNER JOIN
df = pd.merge(left = beatles_songs, right = beatles_lyrics, how = 'inner', left_on = 'title', right_on = 'title')
df.head(10)

We've created a new dataset that contains the song details present in _beatles_songs_ and the song lyrics present in _beatles_lyrics_. 

Now, let us see the number of songs this dataframe possesses.

df.shape

Of the 213 official studio versions of songs, we have data for 211 of them. We can try to find and add the lyrics for the two missing songs but 211 is still good enough for this current project.

## 6. Cleaning the Song Lyrics

Now, let us start cleaning the song lyrics. For this, we start by making a function _preprocess_ which will take each row and then turn each word into lower case, remove all punctuations, drop all numerical characters and then remove any white space left behind after previous removals. The function will then take each word and lemmatize it, provided it is not a stop word, and then join the words to return a clean version of the lyrics.

### CREATING A FUNCTION TO CLEAN SONG LYRICS
stop_words = stopwords.words('english')
lemmatizer = WordNetLemmatizer()
def preprocess(text):
    clean_words = []
    text = text.lower() #turns to lower case
    text = re.sub(r'[^\w\s]+', '', text) #removes punctuation
    text = re.sub('\d+', '', text) #removes numerical characters
    text = text.strip() #removes white spaces
    for word in text.split():
        if word not in stop_words:
            word = lemmatizer.lemmatize(word)
            clean_words.append(word)
    clean_words = ' '.join(clean_words)
    return clean_words

Now let us apply this function to the lyrics of each song.

### APPLYING THE FUNCTION TO LYRICS
df['lyrics'] = df['lyrics'].apply(preprocess)
df.head(10)

We see that all the lyrics have been cleaned.

## 7. Creating a Document Term Matrix

Next, we shall create a Document Term Matrix(DTM) where each row will represent the lyrics of a song and the columns will represent all the words appearing in the lyrics of all the songs. Each _(i,j)_-th entry of this DTM will be the frequency of appearance of the _j_-th word in the lyrics of the _i_-th song. To accomplish this, we shall make a function _get_count_ that will apply _CountVectorizer_ to the lyrics of every song and return a dataframe that is essentially a Document Term Matrix. We shall keep two parameters in mind:

1. ngram_range: the range of consecutive words we want to divide the lyrics into which will serve as features(we've made it a user defined parameter here)
2. max_features: the maximum number of such features

### CREATING A FUNCTION THAT RETURNS A DOCUMENT TERM MATRIX
def get_count(dataframe, m, n):
    vectorizer = CountVectorizer(ngram_range = (m, n), lowercase = False, max_features = 5000)
    count_set = vectorizer.fit_transform(dataframe).todense()
    count_df = pd.DataFrame(count_set, columns = vectorizer.get_feature_names())
    return count_df

Now, let us apply this function to the lyrics of each song by taking only single words as features. We then separate the DTM as X and the song titles as y.

### CREATING A DTM AND RESPONSE VARIABLE
X_lyrics = df['lyrics']
X = get_count(X_lyrics, 1, 1)
y = df['Song']

## 8. Creating a Multinomial Naive Bayes Model

We now construct a Multinomial Naive Bayes model that will predict the song title from a line of lyrics we feed into it. We will set alpha as 1.

### CREATING A MULTINOMIAL NAIVE BAYES CLASSIFIER
clf_nb = MultinomialNB(alpha = 1)
clf_nb.fit(X, y)

## 9. Cleaning the Lyrics to be Predicted

Now, let us write a line of a song that we hope to get the title for.

### LYRICS THAT WE WANT TO FEED INTO THE MODEL
my_lyrics = 'sun is up the sky is blue its beautiful and so are you'

Before feeding this line to the model, it needs to be converted into a DTM. For this, we shall create a function that will do the following:

1. Make a new dataframe _df2_ by dropping all the columns of _df_ except _'Song'_ and _'lyrics'_.
2. Clean the line that we want to feed to our model using the _preprocess_ function.
3. Make a new array with columns _'Song'_ and _'lyrics'_; add NaN and the cleaned lyrics from Step 2. as their respective values.
4. Attach this array to _df2_.
5. Create DTM from the lyrics in _df2_ using the _get_count_ function.
6. Compare the columns of this new DTM with the columns of the DTM we used for training(X) and remove those that are not in X.
7. Finally, return the last row of this new DTM since it has the values for the line we want to feed into the model.

### CREATING A FUNCTION THAT CLEANS THE INPUT LYRICS AND GET IT INTO A DTM FORM
def get_clean_lyrics(lyrics):
    df2 = df.drop(['Core catalogue release(s)', 'Songwriter(s)', 'Lead vocal(s)[d]', 'Year', 'title'], axis = 1) #1
    my_lyrics_cleaned = preprocess(my_lyrics) #2
    df3 = {'Song': np.nan, 'lyrics': my_lyrics_cleaned} #3
    df_new = df2.append(df3, ignore_index = True) #4
    X_new_lyrics = df_new['lyrics']
    X_new = get_count(X_new_lyrics, 1, 1) #5
    remove_features = [] #6
    for word_combo in X_new.columns:
        if word_combo not in X.columns:
            remove_features.append(word_combo)
    X_new.drop(remove_features, axis = 1, inplace = True)
    my_song = X_new.tail(1) #7
    return my_song

Let us check to see if we got the desired result.

### APPLYING THE ABOVE FUNCTION TO THE INPUT LYRICS
my_song = get_clean_lyrics(my_lyrics)
my_song

## 10. Predicting the Song Title

First, let us create a function _get_song_details_ that will return the song title along with the other song details.

### CREATING A FUNCTION THAT RETURNS SONG DETAILS
def get_song_details(song_title):
    print('Song title: %s' % song_title)
    print('Album: %s' % beatles_songs[beatles_songs['Song'] == song_title]['Core catalogue release(s)'].values[0])
    print('Songwriter(s): %s' % beatles_songs[beatles_songs['Song'] == song_title]['Songwriter(s)'].values[0])
    print('Lead vocal(s): %s' % beatles_songs[beatles_songs['Song'] == song_title]['Lead vocal(s)[d]'].values[0])
    print('Year: %d' % beatles_songs[beatles_songs['Song'] == song_title]['Year'])

Now, we try to predict the song title from the lyrics using the Multinomial Naive Bayes model.

### PREDICTING THE SONG TITLE
predicted_song = clf_nb.predict(my_song)[0]
get_song_details(predicted_song)

The model predicts the song title which appears to be correct.
